Cubic CGFloat -> Float 변경

+++ Anything/Module/Cubic.swift
@@ -8,22 +8,22 @@
 
 import UIKit
 
-typealias Easing = (_ t: CGFloat, _ b: CGFloat, _ c: CGFloat, _ d: CGFloat) -> CGFloat
-typealias BackEasing = (_ t: CGFloat, _ b: CGFloat, _ c: CGFloat, _ d: CGFloat, _ s: CGFloat) -> CGFloat
-typealias ElasticEasing = (_ t: CGFloat, _ b: CGFloat, _ c: CGFloat, _ d: CGFloat, _ a: CGFloat, _ p: CGFloat) -> CGFloat
+typealias Easing = (_ t: Float, _ b: Float, _ c: Float, _ d: Float) -> Float
+typealias BackEasing = (_ t: Float, _ b: Float, _ c: Float, _ d: Float, _ s: Float) -> Float
+typealias ElasticEasing = (_ t: Float, _ b: Float, _ c: Float, _ d: Float, _ a: Float, _ p: Float) -> Float
 
 struct Cubic {
-    static var EaseIn: Easing = { (_t, b, c, d) -> CGFloat in
+    static var EaseIn: Easing = { (_t, b, c, d) -> Float in
         let t = _t / d
         return c * t * t * t + b
     }
 
-    static var EaseOut: Easing = { (_t, b, c, d) -> CGFloat in
+    static var EaseOut: Easing = { (_t, b, c, d) -> Float in
         let t = _t / d - 1
         return c * (t * t * t + 1) + b
     }
 
-    static var EaseInOut: Easing = { (_t, b, c, d) -> CGFloat in
+    static var EaseInOut: Easing = { (_t, b, c, d) -> Float in
         var t = _t / (d / 2)
         if t < 1 {
             return c / 2 * t * t * t + b
